@@ -1,4 +1,4 @@
-import React, {Component} from "react";
+import React, {Component, useEffect, useState} from "react";
 import axios, {AxiosError, AxiosPromise, AxiosResponse} from "axios";
 import {RouteComponentProps} from "react-router";
 import {Link} from "react-router-dom";
@@ -18,68 +18,68 @@ interface PackageDetailState {
     reverseDependencies: [] | string;
 }
 
-export default class PackageDetail extends Component<PackageDetailProps & RouteComponentProps,
-    PackageDetailState | {}> {
-    state = {
-        packageName: "",
-        packageDescription: "",
-        dependencies: [],
-        reverseDependencies: []
-    };
+const PackagesDetail = (props: any) => {
+    const [packageName, setPackageName] = useState('');
+    const [packageDescription, setPackageDescription] = useState('');
+    const [dependencies, setDependencies] = useState([]);
+    const [reverseDependencies, setReverseDependencies] = useState([]);
 
-    async componentDidMount() {
-        const {name} = this.props.match.params;
+    const fetchPackages = async () => {
+        const {name} = props.match.params;
         const packageData = await axios.get(`/api/packages/${name}`);
-        packageData.data.Package ? this.setState({packageName: packageData.data.Package}) : "";
-        packageData.data.Description ? this.setState({packageDescription: packageData.data.Description}) : "";
-        packageData.data.Depends ? this.setState({dependencies: packageData.data.Depends}) : [];
-        packageData.data.reverseDepends? this.setState({reverseDependencies: packageData.data.reverseDepends}) : [];
-    }
+        packageData.data.Package ? setPackageName(packageData.data.Package) : "";
+        packageData.data.Description ? setPackageDescription(packageData.data.Description) : "";
+        packageData.data.Depends ? setDependencies(packageData.data.Depends) : [];
+        packageData.data.reverseDepends ? setReverseDependencies(packageData.data.reverseDepends) : [];
+    };
+    useEffect(() => {
+        fetchPackages()
+    }, []);
 
-    render() {
-        return (
-            <div className="ui container">
-                <div className="ui centered card">
-                    <div className="content">
-                        <div className="header">{this.state.packageName}</div>
-                        <div className="description">
-                            <p>{this.state.packageDescription}</p>
-                        </div>
+
+    return (
+        <div className="ui container">
+            <div className="ui centered card">
+                <div className="content">
+                    <div className="header">{packageName}</div>
+                    <div className="description">
+                        <p>{packageDescription}</p>
                     </div>
                 </div>
-                <div className="ui link list">
-                    <h4 style={{textAlign: "center"}} className="ui header">
-                        Dependencies
-                    </h4>
-                    {this.state.dependencies.map((name, index) => (
-                        <Link
-                            to={`/packages/${name}`}
-                            target="_self"
-                            className="item"
-                            style={{textAlign: "center"}}
-                            key={index}
-                        >
-                            {name}
-                        </Link>
-                    ))}
-                </div>
-                <div className="ui link list">
-                    <h4 style={{textAlign: "center"}} className="ui header">
-                        Reverse Dependencies
-                    </h4>
-                    {this.state.reverseDependencies.map((name, index) => (
-                        <Link
-                            to={`/packages/${name}`}
-                            target="_self"
-                            className="item"
-                            style={{textAlign: "center"}}
-                            key={index}
-                        >
-                            {name}
-                        </Link>
-                    ))}
-                </div>
             </div>
-        );
-    }
-}
+            <div className="ui link list">
+                <h4 style={{textAlign: "center"}} className="ui header">
+                    Dependencies
+                </h4>
+                {dependencies.map((name, index) => (
+                    <Link
+                        to={`/packages/${name}`}
+                        target="_self"
+                        className="item"
+                        style={{textAlign: "center"}}
+                        key={index}
+                    >
+                        {name}
+                    </Link>
+                ))}
+            </div>
+            <div className="ui link list">
+                <h4 style={{textAlign: "center"}} className="ui header">
+                    Reverse Dependencies
+                </h4>
+                {reverseDependencies.map((name, index) => (
+                    <Link
+                        to={`/packages/${name}`}
+                        target="_self"
+                        className="item"
+                        style={{textAlign: "center"}}
+                        key={index}
+                    >
+                        {name}
+                    </Link>
+                ))}
+            </div>
+        </div>
+    );
+};
+export default PackagesDetail
