@@ -1,44 +1,45 @@
-import React, { Component } from "react";
-import axios, { AxiosError, AxiosPromise, AxiosResponse } from "axios";
-import { Link } from "react-router-dom";
+import React, { useState, useEffect} from "react";
+import axios, {AxiosError, AxiosPromise, AxiosResponse} from "axios";
+import {Link} from "react-router-dom";
 
 interface PackagesListProps {
-  packagesData: AxiosPromise | AxiosResponse | AxiosError;
-  element: string;
-  index: number;
-  componentDidMount: Promise<void>;
-  renderPackages: () => HTMLAllCollection;
+    packagesData: AxiosPromise | AxiosResponse | AxiosError;
+    element: string;
+    index: number;
+    componentDidMount: Promise<void>;
+    renderPackages: () => HTMLAllCollection;
 }
 
 interface PackagesListState {
-  packages: [] | string[];
+    packages: [] | string[];
 }
 
-export default class PackagesList extends Component<
-  PackagesListProps | {},
-  PackagesListState
-> {
-  state = { packages: [] };
+const PackagesList = () => {
+    const [packages, setPackages] = useState([]);
 
-  async componentDidMount() {
-    const packagesData = await axios.get("/api/packages");
-    packagesData.data ? this.setState({ packages: packagesData.data }) : [];
-  }
+    const fetchPackages = async () => {
+        const packagesData = await axios.get("/api/packages");
+        packagesData.data ? setPackages(packagesData.data) : [];
+    };
 
-  render() {
+    useEffect(() => {
+        fetchPackages()
+    }, []);
+
     return (
-      <div className="ui link list">
-        {this.state.packages.map((name, index) => (
-          <Link
-            to={`/packages/${name}`}
-            className="item"
-            style={{ textAlign: "center" }}
-            key={index}
-          >
-            {name}
-          </Link>
-        ))}
-      </div>
+        <div className="ui link list">
+            {packages.map((name, index) => (
+                <Link
+                    to={`/packages/${name}`}
+                    className="item"
+                    style={{textAlign: "center"}}
+                    key={index}
+                >
+                    {name}
+                </Link>
+            ))}
+        </div>
     );
-  }
-}
+
+};
+export default PackagesList;
